requirementx.txt
python-dotenv
selenium
webdriver-manager
openpyxl
pytest